Remove notnull constraint from generic data model (#9172)
### Motivation and Context

Some databases allow us to generate keys on the server, so when doing an
upsert, a key doesn't have to be specified on the record.
This means that the key on the record should be nullable to enable this
functionliaty.

### Description

- Removing the notnull constraint from the generic data model to support
this scenario.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: